Move external provider modules (fanduel, sleeper, espn, etc) into their own directory for clarity
