+++ backend/src/app.module.ts
@@ -4,14 +4,14 @@ import { DatabaseModule } from '@/infrastructure/database/database.module';
 import { EventsModule } from '@/events/events.module';
 import { LeaguesModule } from '@/leagues/leagues.module';
 import { PlayerStatsModule } from '@/player-stats/player-stats.module';
-import { SleeperModule } from '@/sleeper/sleeper.module';
+import { SleeperModule } from '@/external-providers/sleeper/sleeper.module';
 import { TeamsModule } from '@/teams/teams.module';
 import { UsersModule } from '@/users/users.module';
 import { Module } from '@nestjs/common';
 import { HealthController } from '@/health/health.controller';
 import { ConfigModule } from '@nestjs/config';
-import { FanduelModule } from './fanduel/fanduel.module';
-import { EspnModule } from './espn/espn.module';
+import { FanduelModule } from '@/external-providers/fanduel/fanduel.module';
+import { EspnModule } from '@/external-providers/espn/espn.module';
 
 @Module({
   imports: [
