+++ backend/src/external-providers/fanduel/entities/fanduel-golf.entity.ts
