+++ backend/src/external-providers/fanduel/entities/fanduel.entity.ts
