+++ backend/src/external-providers/sleeper/__fixtures__/sleeper-projections.raw.test.json
