+++ backend/src/external-providers/sleeper/__fixtures__/sleeper-projections.transformed.test.json
