+++ backend/src/external-providers/sleeper/__fixtures__/sleeper-stats.transformed.test.json
