+++ backend/src/infrastructure/test/sdk/functional/sleeper/stats/index.ts
@@ -4,12 +4,12 @@
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
-import type { SleeperStatRequest } from "@/sleeper/entities/sleeper.entity";
+import type { SleeperStatRequest } from "@/external-providers/sleeper/entities/sleeper.entity";
 import type { IConnection } from "@nestia/fetcher";
 import { PlainFetcher } from "@nestia/fetcher/lib/PlainFetcher";
 import type { Primitive, Resolved } from "typia";
 
-import type { SleeperStatResponse } from "../../../../../../sleeper/entities/sleeper.entity";
+import type { SleeperStatResponse } from "@/external-providers/sleeper/entities/sleeper.entity";
 
 /**
  * @controller SleeperController.getStats
