+++ backend/src/player-stats/player-stats.module.ts
@@ -1,9 +1,9 @@
-import { FanduelModule } from '@/fanduel/fanduel.module';
-import { SleeperModule } from '@/sleeper/sleeper.module';
 import { EventsModule } from '@/events/events.module';
+import { FanduelModule } from '@/external-providers/fanduel/fanduel.module';
+import { SleeperModule } from '@/external-providers/sleeper/sleeper.module';
 import { Module } from '@nestjs/common';
-import { PlayerStatsService } from './player-stats.service';
 import { PlayerStatsController } from './player-stats.controller';
+import { PlayerStatsService } from './player-stats.service';
 
 @Module({
   imports: [SleeperModule, FanduelModule, EventsModule],
