+++ backend/src/teams/teams.module.ts
@@ -1,17 +1,13 @@
+import { EventsModule } from '@/events/events.module';
+import { db } from '@/infrastructure/database/database';
+import { DatabaseModule } from '@/infrastructure/database/database.module';
 import { LeaguesModule } from '@/leagues/leagues.module';
 import { PlayerStatsModule } from '@/player-stats/player-stats.module';
-import { SleeperModule } from '@/sleeper/sleeper.module';
-import {
-  DatabaseTeamsEntryRepository,
-  TeamsEntryRepository,
-} from '@/teams/teams-entry.repository';
+import { DatabaseTeamsEntryRepository, TeamsEntryRepository } from '@/teams/teams-entry.repository';
 import { Module } from '@nestjs/common';
-import { TeamsService } from './teams.service';
 import { TeamsController } from './teams.controller';
-import { db } from '@/infrastructure/database/database';
 import { DatabaseTeamsRepository, TeamsRepository } from './teams.repository';
-import { DatabaseModule } from '@/infrastructure/database/database.module';
-import { EventsModule } from '@/events/events.module';
+import { TeamsService } from './teams.service';
 
 @Module({
   controllers: [TeamsController],
@@ -28,7 +24,7 @@ import { EventsModule } from '@/events/events.module';
     {
       provide: TeamsEntryRepository,
       useClass: DatabaseTeamsEntryRepository,
-    }
+    },
   ],
   imports: [DatabaseModule, LeaguesModule, PlayerStatsModule, EventsModule],
 })
