+++ backend/src/users/users.module.ts
@@ -1,9 +1,9 @@
-import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-import { DatabaseUsersRepository, UsersRepository } from './users.repository';
 import { db } from '@/infrastructure/database/database';
 import { DatabaseModule } from '@/infrastructure/database/database.module';
+import { Module } from '@nestjs/common';
+import { UsersController } from './users.controller';
+import { DatabaseUsersRepository, UsersRepository } from './users.repository';
+import { UsersService } from './users.service';
 
 @Module({
   controllers: [UsersController],
